requirements.txt
# THE AIRLOCK v5.1.1 — Core Dependencies (pinned)
#
# Kurulum:
#   pip install -r requirements.txt
#
# Bu dosya scripts/setup.sh ADIM 9 tarafından kullanılır.
# Donanım bağımlılıkları için: requirements-hardware.txt
PyYAML==6.0.2
python-magic==0.4.27
yara-python==4.5.1
PyNaCl==1.5.0
psutil==6.1.0
pyudev==0.24.3
pyclamd==0.4.0
chardet==5.2.0
watchdog==6.0.0
Pillow==11.0.0
